TEST: import coreCommand again, as the unexplained load time dropped dramatically
There is a noticeable delay in `actions/checkout@v3` after this line:

	▶ Run dscho/checkout@main

The delay is diagnosed by instrumenting the compiled Action: instead of
starting the core logic we throw an exception.

With an only lightly instrumented version, we see an almost two minute
long delay after that line:
https://github.com/dscho/git-for-windows-automation/actions/runs/4321187056/jobs/7542820362#step:2:49

Removing some imports at first did not change that, but when removing
all but the first import, the delay dropped to two seconds:
https://github.com/dscho/git-for-windows-automation/actions/runs/4321187056/jobs/7545290126#step:2:49

Let's see whether loading `coreCommand` is what causes that surprising
difference.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>