ci: verify that an existing sparse checkout can be made unsparse
There are use cases in particular with non-ephemeral runners where an
existing worktree (that has been initialized as a sparse checkout) is
reused in subsequent CI runs (where `actions/checkout` is run _without_
any `sparse-checkout` parameter).

In these scenarios, we need to make sure that the sparse checkout is
disabled before checking out the files.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>